refactor(dashboard): renamed tabs and views
