+++ src/components/auth/GameManagement.jsx
@@ -20,7 +20,6 @@ import dayjs from "dayjs";
 import TableData from "../common/TableData";
 import PhotoLibraryIcon from "@mui/icons-material/PhotoLibrary";
 import CloseIcon from "@mui/icons-material/Close";
-import Heading from "../common/Heading";
 import { useTheme } from "@mui/material/styles";
 import FormTextField from "../common/FormTextField";
 import FormSelect from "../common/FormSelect";
